Personally was experiencing buggy behaviour with always on top, removed the ability for hover to put things on top unless in always on top mode.
Some may like this, some not, but I found it was near impossible to use third party apps with the hover, as it would randomly put things on top even if not hovered.  This will prevent this, however it removes the hover to front if one isn't in Always on top, worth it for me personally.